error handling not inside function

+++ savintsev.ilya/P4/dynamicstrings.h
@@ -7,7 +7,7 @@ namespace savintsev
   char * inputEndlessStr(std::istream & in);
   char * increaseStrSize(char * old, size_t m);
   char * allocMemoryStr(size_t capacity);
-  int getNumIdenticalInRow(char * c);
+  size_t getNumIdenticalInRow(char * c);
 }
 
 #endif
